requirements.txt
streamlit>=1.36
python-dotenv>=1.0
openai>=1.0
elevenlabs>=1.0
plotly>=5.0
httpx>=0.27
